+++ service/position.go
@@ -2,13 +2,16 @@ package service
 
 import (
 	"context"
+	"fmt"
 
 	"code.vegaprotocol.io/data-node/entities"
 	"code.vegaprotocol.io/data-node/logging"
 	"code.vegaprotocol.io/data-node/utils"
+	lru "github.com/hashicorp/golang-lru"
 )
 
-type positionStore interface {
+//go:generate go run github.com/golang/mock/mockgen -destination mocks/position_mock.go -package mocks code.vegaprotocol.io/data-node/service PositionStore
+type PositionStore interface {
 	Flush(ctx context.Context) ([]entities.Position, error)
 	Add(ctx context.Context, p entities.Position) error
 	GetByMarketAndParty(ctx context.Context, marketID entities.MarketID, partyID entities.PartyID) (entities.Position, error)
@@ -17,17 +20,27 @@ type positionStore interface {
 	GetAll(ctx context.Context) ([]entities.Position, error)
 }
 
+type positionCacheKey struct {
+	MarketID entities.MarketID
+	PartyID  entities.PartyID
+}
 type Position struct {
 	log      *logging.Logger
-	store    positionStore
+	store    PositionStore
 	observer utils.Observer[entities.Position]
+	cache    *lru.Cache
 }
 
-func NewPosition(store positionStore, log *logging.Logger) *Position {
+func NewPosition(store PositionStore, log *logging.Logger) *Position {
+	cache, err := lru.New(10000)
+	if err != nil {
+		panic(err)
+	}
 	return &Position{
 		store:    store,
 		log:      log,
 		observer: utils.NewObserver[entities.Position]("positions", log, 0, 0),
+		cache:    cache,
 	}
 }
 
@@ -41,11 +54,33 @@ func (p *Position) Flush(ctx context.Context) error {
 }
 
 func (p *Position) Add(ctx context.Context, pos entities.Position) error {
+	key := positionCacheKey{pos.MarketID, pos.PartyID}
+	p.cache.Add(key, pos)
 	return p.store.Add(ctx, pos)
 }
 
 func (p *Position) GetByMarketAndParty(ctx context.Context, marketID entities.MarketID, partyID entities.PartyID) (entities.Position, error) {
-	return p.store.GetByMarketAndParty(ctx, marketID, partyID)
+	key := positionCacheKey{marketID, partyID}
+	value, ok := p.cache.Get(key)
+	if !ok {
+		pos, err := p.store.GetByMarketAndParty(ctx, marketID, partyID)
+		if err == nil {
+			p.cache.Add(key, pos)
+		} else { // If store errors in the cache too
+			p.cache.Add(key, err)
+		}
+
+		return pos, err
+	}
+
+	switch v := value.(type) {
+	case entities.Position:
+		return v, nil
+	case error:
+		return entities.Position{}, v
+	default:
+		return entities.Position{}, fmt.Errorf("unknown type in cache")
+	}
 }
 
 func (p *Position) GetByMarket(ctx context.Context, marketID entities.MarketID) ([]entities.Position, error) {